collaboration_system/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
redis-om==0.3.0

# WebSocket
websockets==12.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-decouple==3.8

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Performance
uvloop==0.19.0
orjson==3.9.10

# Optional: AI Integration (for advanced features)
# openai==1.3.7
# langchain==0.0.340
# transformers==4.35.2
# torch==2.1.1

# Optional: Document Processing
# beautifulsoup4==4.12.2
# markdown==3.5.1
# PyPDF2==3.0.1

# Optional: Search Integration
# elasticsearch==8.11.0
# whoosh==2.7.4